separating out utils and actually reverting existing data logic
