+++ language/metrics-output-plugin/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./metrics-output";
+export * from "./utils";
